Sprite Iso 639
This is a sprite to use images of the basic languages ​​covered by the
ISO 639 standard in a simple way. Each flag has a height of 33px and a
width of 50px. To use a banner add class = "language language-xx" (where
xx is the language code in ISO 639) to the corresponding tag